+++ tests/sync/server.py
@@ -1,19 +1,22 @@
 import contextlib
 import ssl
 import threading
+import urllib.parse
 
+from websockets.sync.router import *
 from websockets.sync.server import *
 
 
-def get_uri(server):
-    secure = isinstance(server.socket, ssl.SSLSocket)  # hack
+def get_uri(server, secure=None):
+    if secure is None:
+        secure = isinstance(server.socket, ssl.SSLSocket)  # hack
     protocol = "wss" if secure else "ws"
     host, port = server.socket.getsockname()
     return f"{protocol}://{host}:{port}"
 
 
 def handler(ws):
-    path = ws.request.path
+    path = urllib.parse.urlparse(ws.request.path).path
     if path == "/":
         # The default path is an eval shell.
         for expr in ws:
@@ -34,8 +37,14 @@ class EvalShellMixin:
 
 
 @contextlib.contextmanager
-def run_server(handler=handler, host="localhost", port=0, **kwargs):
-    with serve(handler, host, port, **kwargs) as server:
+def run_server_or_router(
+    serve_or_route,
+    handler_or_url_map,
+    host="localhost",
+    port=0,
+    **kwargs,
+):
+    with serve_or_route(handler_or_url_map, host, port, **kwargs) as server:
         thread = threading.Thread(target=server.serve_forever)
         thread.start()
 
@@ -63,9 +72,22 @@ def run_server(handler=handler, host="localhost", port=0, **kwargs):
                 handler_thread.join()
 
 
+def run_server(handler=handler, **kwargs):
+    return run_server_or_router(serve, handler, **kwargs)
+
+
+def run_router(url_map, **kwargs):
+    return run_server_or_router(route, url_map, **kwargs)
+
+
 @contextlib.contextmanager
-def run_unix_server(path, handler=handler, **kwargs):
-    with unix_serve(handler, path, **kwargs) as server:
+def run_unix_server_or_router(
+    path,
+    unix_serve_or_route,
+    handler_or_url_map,
+    **kwargs,
+):
+    with unix_serve_or_route(handler_or_url_map, path, **kwargs) as server:
         thread = threading.Thread(target=server.serve_forever)
         thread.start()
         try:
@@ -73,3 +95,11 @@ def run_unix_server(path, handler=handler, **kwargs):
         finally:
             server.shutdown()
             thread.join()
+
+
+def run_unix_server(path, handler=handler, **kwargs):
+    return run_unix_server_or_router(path, unix_serve, handler, **kwargs)
+
+
+def run_unix_router(path, url_map, **kwargs):
+    return run_unix_server_or_router(path, unix_route, url_map, **kwargs)